feat: 현재 채널 라이브 및 동영상 목록 조회 /코드 동작 테스트 완

+++ Mytube/src/main/java/com/flab/Mytube/dto/movie/response/JoinChatResponse.java
@@ -8,7 +8,7 @@ import java.math.BigInteger;
 @Getter
 @AllArgsConstructor
 public class JoinChatResponse {
-    private BigInteger id;
+    private long id;
     private int code;
     private String message;
 }
